+++ src/major-node-version.js
@@ -1,0 +1,1 @@
+module.exports = parseInt(process.version.split('.')[0].match(/\d+/)[0])