+++ test/workers/es6-module.mjs
@@ -1,0 +1,7 @@
+export function fn (arg1) {
+  return 'got ' + arg1
+}
+
+// export default function fn2 (arg1) {
+//   return 'default ' + arg1
+// }